add all tests into one MainTest class
