+++ .github/workflows/test-actions.yml
@@ -14,20 +14,8 @@ jobs:
           cache: maven
       - name: Compile unit tests with Maven
         run: mvn test-compile
-      - name: Run unit test with Maven
+      - name: Run unit and integration test with Maven
         run: mvn -X -e -Dtest=MainTest test
-      #      - name: Compile project
-      #        run: javac -cp $(mvn -q dependency:build-classpath -Dmdep.outputFile=/dev/fd/1) src/main/java/org/example/Main.java
-      - name: Run project and integration tests
-        run: |
-          mvn clean package
-          java -cp target/your-artifact.jar org.example.Main &
-          sleep 10 
-          mvn -Dtest=SparkIntegrationTest test
-      #      - name: run project
-      #        run: mvn exec:java -Dexec.mainClass=org.example.Main &
-      #      - name: run integration tests
-      #        run: mvn -X -e -Dtest=SparkIntegrationTest test
       - name: exit from console
         run: exit
